requirements.txt
langchain
langchain-aws
langchain-openai
langchain-google-community
python-dotenv
